+++ core/server/api_container/server/service_network/user_service_launcher/user_service_launcher.go
@@ -17,10 +17,6 @@ import (
 	"net"
 )
 
-const (
-	enclaveDataDirMntDirpath = "/kurtosis-enclave-data"
-)
-
 /*
 Convenience struct whose only purpose is launching user services
 */
@@ -28,11 +24,10 @@ type UserServiceLauncher struct {
 	kurtosisBackend          backend_interface.KurtosisBackend
 	filesArtifactExpander    *files_artifact_expander.FilesArtifactExpander
 	freeIpAddrTracker        *lib.FreeIpAddrTracker
-	enclaveDataDirpathOnHostMachine string
 }
 
-func NewUserServiceLauncher(kurtosisBackend backend_interface.KurtosisBackend, filesArtifactExpander *files_artifact_expander.FilesArtifactExpander, freeIpAddrTracker *lib.FreeIpAddrTracker, enclaveDataDirpathOnHostMachine string) *UserServiceLauncher {
-	return &UserServiceLauncher{kurtosisBackend: kurtosisBackend, filesArtifactExpander: filesArtifactExpander, freeIpAddrTracker: freeIpAddrTracker, enclaveDataDirpathOnHostMachine: enclaveDataDirpathOnHostMachine}
+func NewUserServiceLauncher(kurtosisBackend backend_interface.KurtosisBackend, filesArtifactExpander *files_artifact_expander.FilesArtifactExpander, freeIpAddrTracker *lib.FreeIpAddrTracker) *UserServiceLauncher {
+	return &UserServiceLauncher{kurtosisBackend: kurtosisBackend, filesArtifactExpander: filesArtifactExpander, freeIpAddrTracker: freeIpAddrTracker}
 }
 
 /**
@@ -98,8 +93,6 @@ func (launcher UserServiceLauncher) Launch(
 		entrypointArgs,
 		cmdArgs,
 		envVars,
-		launcher.enclaveDataDirpathOnHostMachine,
-		enclaveDataDirMntDirpath,
 		artifactVolumeMounts,
 	)
 	if err != nil {